[GDE] Modification of runnable read_distance

+++ Ultrasonic_Sensor_Driver_read_distance.c
@@ -1,7 +1,7 @@
 #include "Fmi_Cfg.h" // Ids for <XXX>WriteToSharedObj
 #include "ASimSharedObjUtils.h" // <XXX>WriteToSharedObj
 
-void read_distance(int* distance){
+void read_distance(uint16_t* distance){
   fmi2Integer data;
   
   fmi2IntegerReadFromWriteBuffer(InputId_Ultrasonic_Sensor_Driver_read_distance_0, &data);
